+++ src/Codec/FFmpeg/Internal/Linear.hs
@@ -1,6 +1,6 @@
 {-# LANGUAGE ScopedTypeVariables #-}
--- | A basic three component vector.
-module Codec.FFmpeg.Internal.V3 where
+-- | Minimal operations on small vector types.
+module Codec.FFmpeg.Internal.Linear where
 import Control.Applicative
 import Foreign.C.Types
 import Foreign.Ptr (castPtr)
@@ -28,3 +28,20 @@ qd (V3 x1 y1 z1) (V3 x2 y2 z2) = let dx = x2 - x1
                                      dy = y2 - y1
                                      dz = z2 - z1
                                  in dx * dx + dy * dy + dz * dz
+
+data V4 a = V4 !a !a !a !a
+
+instance Functor V4 where
+  fmap f (V4 x y z w) = V4 (f x) (f y) (f z) (f w)
+
+instance Storable a => Storable (V4 a) where
+  sizeOf _ = 4 * sizeOf (undefined::a)
+  alignment _ = alignment (undefined::a)
+  peek ptr = let ptr' = castPtr ptr
+             in V4 <$> peek ptr' <*> peekElemOff ptr' 1
+                   <*> peekElemOff ptr' 2 <*> peekElemOff ptr' 3
+  poke ptr (V4 x y z w) = let ptr' = castPtr ptr
+                          in do poke ptr' x
+                                pokeElemOff ptr' 1 y
+                                pokeElemOff ptr' 2 z
+                                pokeElemOff ptr' 3 w